Crate strymon_rpc

This crate defines the remote-procedure call protocol used between the different components in Strymon Core.

All types here are based on the framework outlined in strymon_communication::rpc. Each submodule contains the type definitions for one interface. The coordinator module contains all requests sent to the coordinator by various components, while the executor contains the messages sent from the coordinator to a registered executor.

Modules

coordinator
This module contains all requests (and the respones) to the the coordinator by the strymon submit tool, submitted and running jobs, as well as connecting executors.

executor
The interface exposed by executors. This interface is currently only intended to be used by the coordinator to send requests to a registered executor.