Enum strymon_rpc::executor::SpawnError

pub enum SpawnError {
    InvalidRequest,
    FileNotFound,
    WorkdirCreationFailed,
    FetchFailed,
    ExecFailed,
}

The error message returned by the executor for failed spawn requests.

Variants
InvalidRequest

The request was invalid.

FileNotFound

The executor was unable to locate the downloaded file.

WorkdirCreationFailed

The working directory for the job could not be created.

FetchFailed

The executable was unable to fetch the executable file.

ExecFailed

Launching the executable failed.